q2_document_search/requirements.txt
streamlit==1.28.1
torch==2.1.0
transformers==4.35.2
sentence-transformers==2.2.2
pdfplumber==0.10.2
python-docx==1.1.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
spacy==3.7.2
nltk==3.8.1
plotly==5.17.0
streamlit-option-menu==0.3.6
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0 